# tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CausaOficial": {
        "description": "Las nueve causas de variación que enumera literalmente la ficha.\n\n``TipoMovimiento`` es el vocabulario técnico del CRM; ``CausaOficial`` es el\nvocabulario del enunciado y de la evaluación. Nunca se mezclan: se traducen con\n``causa_oficial_de``.",
        "enum": [
          "CAMBIO_DE_PLAN",
          "EQUIPO_FINANCIADO",
          "COMPRA_DE_PAQUETES",
          "CARGOS_ADICIONALES",
          "PROMOCIONES_VENCIDAS",
          "NOTAS_CREDITO_DEBITO",
          "PRORRATEOS",
          "RECONEXIONES",
          "AJUSTES_POR_DIAS_DE_SUSPENSION"
        ],
        "title": "CausaOficial",
        "type": "string"
      },
...
      "FamiliaConcepto": {
        "description": "Naturaleza contable del concepto; decide si se prorratea y cómo se narra.",
        "enum": [
          "RECURRENTE",
          "UNICO",
          "AJUSTE",
          "FINANCIAMIENTO",
          "IMPUESTO",
          "CREDITO"
        ],
        "title": "FamiliaConcepto",
        "type": "string"
      },
      "FichaConcepto": {
        "additionalProperties": false,
        "description": "Ficha de un concepto tal como se le muestra al cliente.",
        "properties": {
          "afecto_igv": {
            "default": true,
            "title": "Afecto Igv",
            "type": "boolean"
          },
          "causa_oficial": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CausaOficial"
              },
              {
                "type": "null"
              }
            ]
          },
          "causas_permitidas": {
            "items": {
              "type": "string"
            },
            "title": "Causas Permitidas",
            "type": "array"
          },
          "cifras_retiradas": {
            "description": "Marcadores que el saneador puso en lugar de las cifras del corpus",
            "items": {
              "type": "string"
            },
            "title": "Cifras Retiradas",
            "type": "array"
          },
          "concepto_id": {
            "title": "Concepto Id",
            "type": "string"
          },
          "definicion_cliente": {
            "title": "Definicion Cliente",
            "type": "string"
          },
          "detalle_rag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Texto ampliado del corpus, ya saneado de cifras",
            "title": "Detalle Rag"
          },
          "ejemplo_variacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ejemplo Variacion"
          },
          "familia": {
            "$ref": "#/components/schemas/FamiliaConcepto"
          },
          "nombre_comercial": {
            "title": "Nombre Comercial",
            "type": "string"
          },
          "prorrateable": {
            "default": false,
            "title": "Prorrateable",
            "type": "boolean"
          },
          "sinonimos": {
            "items": {
              "type": "string"
            },
            "title": "Sinonimos",
            "type": "array"
          }
        },
        "required": [
          "concepto_id",
          "nombre_comercial",
          "familia",
          "definicion_cliente"
        ],
        "title": "FichaConcepto",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResumenConcepto": {
        "additionalProperties": false,
        "description": "Entrada del listado del catálogo.",
        "properties": {
          "causa_oficial": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CausaOficial"
              },
              {
                "type": "null"
              }
            ]
          },
          "concepto_id": {
            "title": "Concepto Id",
            "type": "string"
          },
          "familia": {
            "$ref": "#/components/schemas/FamiliaConcepto"
          },
          "nombre_comercial": {
            "title": "Nombre Comercial",
            "type": "string"
          }
        },
        "required": [
          "concepto_id",
          "nombre_comercial",
          "familia"
        ],
        "title": "ResumenConcepto",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT HS256 emitido por /dev/token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "concursos-cis@ulima.edu.pe",
      "name": "Equipo recibo-claro"
    },
...
    "license": {
      "name": "MIT"
    },
    "summary": "Explica la variación del recibo con cero invenciones financieras, verificadas en código y comprobables por logs.",
    "title": "recibo-claro · Explicación inteligente de recibos Movistar",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auditoria": {
      "get": {
        "description": "Devuelve los eventos del turno, su resumen y si la cadena está íntegra.",
        "operationId": "obtener_auditoria_v1_auditoria_get",
        "parameters": [
          {
            "description": "Traza devuelta por /v1/explicar",
            "in": "query",
            "name": "trace_id",
            "required": true,
            "schema": {
              "description": "Traza devuelta por /v1/explicar",
              "title": "Trace Id",
              "type": "string"
            }
          },
          {
            "description": "Adjunta el detalle completo",
            "in": "query",
            "name": "incluir_eventos",
            "required": false,
            "schema": {
              "default": true,
              "description": "Adjunta el detalle completo",
              "title": "Incluir Eventos",
              "type": "boolean"
            }
          },
          {
            "description": "Filtra por etapas separadas por coma, p. ej. VERIFY,RESPONSE",
            "in": "query",
            "name": "etapas",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtra por etapas separadas por coma, p. ej. VERIFY,RESPONSE",
              "title": "Etapas"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Obtener Auditoria V1 Auditoria Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "TRAZA_NO_ENCONTRADA"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Eventos auditados de un turno y validez de la cadena de hashes",
        "tags": [
          "auditoria"
        ]
      }
    },
    "/v1/auditoria/cadena": {
      "get": {
        "description": "Recorre toda la bitácora y dice si algún evento fue alterado.\n\nNo devuelve contenido de ningún turno: solo el veredicto y, si falla, el índice del\nprimer evento cuyo hash no cuadra.",
        "operationId": "verificar_cadena_completa_v1_auditoria_cadena_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Verificar Cadena Completa V1 Auditoria Cadena Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Verifica la cadena de hashes completa",
        "tags": [
          "auditoria"
        ]
      }
    },
    "/v1/catalogo": {
      "get": {
        "description": "Devuelve el catálogo de conceptos que se le pueden nombrar a un cliente.",
        "operationId": "listar_v1_catalogo_get",
        "parameters": [
          {
            "description": "Filtra por familia",
            "in": "query",
            "name": "familia",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FamiliaConcepto"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtra por familia",
              "title": "Familia"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResumenConcepto"
                  },
                  "title": "Response Listar V1 Catalogo Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Lista de conceptos visibles para el cliente",
        "tags": [
          "catalogo"
        ]
      }
    },
    "/v1/catalogo/{concepto_id}": {
      "get": {
        "description": "Explica un concepto del recibo sin exponer ningún dato de cuenta.\n\nUn ``concepto_id`` desconocido es un 404, no una explicación inventada: en la\nconversación ese mismo caso dispara la regla dura *\"concepto fuera de catálogo\"* y\nderiva a un asesor.",
        "operationId": "obtener_concepto_v1_catalogo__concepto_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "concepto_id",
            "required": true,
            "schema": {
              "title": "Concepto Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FichaConcepto"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "CONCEPTO_NO_ENCONTRADO"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Definición de un concepto en lenguaje de cliente",
        "tags": [
          "catalogo"
        ]
      }
    },
    "/v1/catalogo/{concepto_id}/crudo": {
      "get": {
        "description": "Vuelca la entrada de ``rules.yaml`` del concepto, incluida su parte técnica.\n\nSirve para auditar qué causas admite cada concepto (``regla_concepto_causa``) sin\nabrir el fichero de reglas en el servidor.",
        "operationId": "obtener_concepto_crudo_v1_catalogo__concepto_id__crudo_get",
        "parameters": [
          {
            "in": "path",
            "name": "concepto_id",
            "required": true,
            "schema": {
              "title": "Concepto Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Obtener Concepto Crudo V1 Catalogo  Concepto Id  Crudo Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ficha completa del concepto (diagnóstico)",
        "tags": [
          "catalogo"
        ]
      }
    },
...
  },
...
}